+++ tests/testthat/test-breaks.R
@@ -81,6 +81,16 @@ test_that("brk_evenly", {
 })
 
 
+test_that("brk_proportions", {
+  b <- brk_res(brk_proportions(c(0.2, 0.8)), 0:10)
+  expect_identical(as.vector(b), c(2, 8))
+
+  expect_error(brk_proportions(c(0, 1, 2)))
+  expect_error(brk_proportions(c(-1, 0.5)))
+  expect_error(brk_proportions(c(0.5, NA)))
+})
+
+
 test_that("brk_mean_sd", {
   x <- rnorm(100)
   expect_silent(b <- brk_res(brk_mean_sd(1:3), x = x))
@@ -122,6 +132,18 @@ test_that("brk_equally", {
 
   brks <- brk_res(brk_equally(3))
   expect_equivalent(brks, brk_res(brk_quantiles(0:3/3)))
+})
+
+
+test_that("brk_pretty", {
+  expect_silent(brks <- brk_res(brk_pretty(5), x = 1:10))
+  expect_equivalent(brks, brk_res(brk_default(pretty(1:10)), x = 1:10))
+
+  expect_silent(brks2 <- brk_res(brk_pretty(5, high.u.bias = 0), x = 1:10))
+  expect_equivalent(
+    brks2,
+    brk_res(brk_default(pretty(1:10, high.u.bias = 0)), x = 1:10)
+  )
 })